+++ packages/presets/index.js
@@ -1,61 +1,17 @@
-export {
-  Pres, Pres as Preset, Presm, demo, randPres
-} from '@palett/pres'
 import {
-  AQUA, ATLAS, AURORA, AZURE, FRESH, INSTA, JUNGLE, LAVA, METRO, MOSS, OCEAN, PLANET, POME, SUBTLE, VIOLA
+  AQUA, ATLAS, AURORA, AZURE, FRESH, INSTA, JUNGLE, LAVA, METRO, MOSS, OCEAN, PLANET, POME, SUBTLE, VIOLA,
 } from './static/material.js'
 import {
   AFRO, ARUBA, AZALE, BERING, BESQUE, BISTRO, BLUSH, BRANDY, DECANTE, DUSKY, ENSIGN, KELLY, LILAC, NORSE, PAGODA,
-  PERSIAN, PINE, PRETTY, ROCOCCO, RODD, SANDY, SUMMER, TOBACCO, WINE
+  PERSIAN, PINE, PRETTY, ROCOCCO, RODD, SANDY, SUMMER, TOBACCO, TROPIC, WINE,
 } from './static/pavtone.js'
 
-export const Presets = {
-  aqua: AQUA,
-  atlas: ATLAS,
-  aurora: AURORA,
-  azure: AZURE,
-  moss: MOSS,
-  jungle: JUNGLE,
-  ocean: OCEAN,
-  lava: LAVA,
-  planet: PLANET,
-  metro: METRO,
-  subtle: SUBTLE,
-  insta: INSTA,
-  pome: POME,
-  viola: VIOLA,
-  besque: BESQUE,
-  fresh: FRESH,
-  rodd: RODD,
-  azale: AZALE,
-  afro: AFRO,
-  summer: SUMMER,
-  sandy: SANDY,
-  tobacco: TOBACCO,
-  brandy: BRANDY,
-  decante: DECANTE,
-  bistro: BISTRO,
-  kelly: KELLY,
-  pagoda: PAGODA,
-  pine: PINE,
-  norse: NORSE,
-  aruba: ARUBA,
-  bering: BERING,
-  dusky: DUSKY,
-  ensign: ENSIGN,
-  persian: PERSIAN,
-  lilac: LILAC,
-  wine: WINE,
-  rococco: ROCOCCO,
-  pretty: PRETTY,
-  blush: BLUSH
-}
-
 export {
-  AQUA, ATLAS, AURORA, AZURE, FRESH, INSTA, JUNGLE, LAVA,
-  METRO, MOSS, OCEAN, PLANET, POME, SUBTLE, VIOLA
+  AQUA, ATLAS, AURORA, AZURE, FRESH, INSTA, JUNGLE, LAVA, METRO, MOSS, OCEAN, PLANET, POME, SUBTLE, VIOLA,
 } from './static/material.js'
 export {
-  AFRO, ARUBA, AZALE, BERING, BESQUE, BISTRO, BLUSH, BRANDY, DECANTE, DUSKY, ENSIGN, KELLY,
-  LILAC, NORSE, PAGODA, PERSIAN, PINE, PRETTY, ROCOCCO, RODD, SANDY, SUMMER, TOBACCO, WINE
+  AFRO, ARUBA, AZALE, BERING, BESQUE, BISTRO, BLUSH, BRANDY, DECANTE, DUSKY, ENSIGN, KELLY, LILAC, NORSE, PAGODA,
+  PERSIAN, PINE, PRETTY, ROCOCCO, RODD, SANDY, SUMMER, TOBACCO, TROPIC, WINE,
 } from './static/pavtone.js'
+
+export { Pres, Presm } from '@palett/pres'