+++ packages/convert/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @palett/convert
 
+## 0.12.15
+
+### Patch Changes
+
+- minor fix
+
 ## 0.12.14
 
 ### Patch Changes