+++ packages/fluo-vector/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @palett/fluo-vector
 
+## 0.12.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/fluo@0.12.15
+
 ## 0.12.14
 
 ### Patch Changes