+++ packages/munsell/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @palett/munsell
 
+## 0.12.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/convert@0.12.15
+
 ## 0.12.14
 
 ### Patch Changes