+++ packages/presets/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @palett/presets
 
+## 0.12.15
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/cards@0.12.15
+  - @palett/pres@0.12.15
+
 ## 0.12.14
 
 ### Patch Changes